backend/pyproject.toml
[project]
name = "saimon-ia-backend"
version = "0.1.0"
description = "Backend API para SAIMON IA"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.23",
  "pydantic>=2.6",
  "python-dotenv>=1.0",
  "httpx>=0.26",
]

[tool.uvicorn]
factory = false
